+++ lua/kustomize/validate.lua
@@ -20,34 +20,21 @@ M.run_validation = function(config, fileToValidate)
 end
 
 M.validate = function(config)
-  -- TODO: move to init.lua?
-  if not utils.is_module_available("plenary") then
-    utils.error("Could not load https://github.com/nvim-lua/plenary.nvim")
-  end
   if not utils.check_exec("kubeconform") then
     utils.error("kubeconform" .. " was not found on your path")
     return
   end
   local fileToValidate
-  local t = os.tmpname()
-  local tmpFile = t .. ".yaml"
+  local tmpFile = utils.create_temp_file_string("yaml")
   local bufName = vim.api.nvim_buf_get_name(0)
-  if path:new(bufName):is_file() then
+  local is_file = path:new(bufName):is_file()
+  if is_file then
     -- file exists on disk
     fileToValidate = path:new({ bufName, sep = utils.path_separator() }):absolute()
   else
     -- if buffer content is not file on disk
-    local bufferData = vim.api.nvim_buf_get_text(0, 0, 0, -1, -1, {})
     fileToValidate = tmpFile
-    local f = io.open(tmpFile, "w+b")
-    if f ~= nil then
-      local data = table.concat(bufferData, "\n")
-      f:write(data)
-      f:flush()
-      if not f:close() then
-        utils.warn("temporary file handler could not be closed")
-      end
-    end
+    utils.create_file_from_current_buffer_content(tmpFile)
   end
   local err, out = M.run_validation(config, fileToValidate)
   if next(err) ~= nil then
@@ -59,10 +46,9 @@ M.validate = function(config)
   elseif next(out) == nil then
     utils.info("no issues found")
   end
-  -- can create an empty file http://www.lua.org/manual/5.1/manual.html#pdf-os.tmpname
-  -- TODO: error handlin?
-  os.remove(t)
-  os.remove(tmpFile)
+  if not is_file then
+    utils.delete_file(tmpFile)
+  end
 end
 
 return M