+++ CHANGELOG.md
@@ -5,6 +5,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.8.0](https://github.com/Allaman/kustomize.nvim/compare/v4.7.0...v4.8.0) (2024-03-02)
+
+
+### Features
+
+* Add LuaSnip integration ([795e2b8](https://github.com/Allaman/kustomize.nvim/commit/795e2b89e8ce4798b78ff97ba093dc780c533413))
+
+## [4.7.0](https://github.com/Allaman/kustomize.nvim/compare/v4.6.0...v4.7.0) (2024-02-03)
+
+
+### Features
+
+* Add exclude_patterns to kind list [#53](https://github.com/Allaman/kustomize.nvim/issues/53) ([58f8b2a](https://github.com/Allaman/kustomize.nvim/commit/58f8b2a4ddda24a7bf8602ac679e08751643502f))
+
+## [4.6.0](https://github.com/Allaman/kustomize.nvim/compare/v4.5.2...v4.6.0) (2024-02-03)
+
+
+### Features
+
+* Add args for kinds and deprecations [#51](https://github.com/Allaman/kustomize.nvim/issues/51) ([dd5bf66](https://github.com/Allaman/kustomize.nvim/commit/dd5bf66f54c139cfd9801234a70d78a773432f31))
+
+
+### Bug Fixes
+
+* Deprecated call ([0c4ee06](https://github.com/Allaman/kustomize.nvim/commit/0c4ee060ef8991a903bd9cf9d40187804da7fd78))
+
+## [4.5.2](https://github.com/Allaman/kustomize.nvim/compare/v4.5.1...v4.5.2) (2023-12-22)
+
+
+### Bug Fixes
+
+* Add error handling to os.remove calls ([#47](https://github.com/Allaman/kustomize.nvim/issues/47)) ([ccd321f](https://github.com/Allaman/kustomize.nvim/commit/ccd321f31e19c147b11716e5b25a54fdd126dd23))
+
 ## [4.5.1](https://github.com/Allaman/kustomize.nvim/compare/v4.5.0...v4.5.1) (2023-12-16)